feat!: update to MDEx 0.10
MDEx 0.10 has a few breaking changes which should have minimal impact on
Tableau, but the larger change is that there is explicit support for
plugins in MDEx, using the similar patterns as Req does.

To simplify the use of plugins in Tableau, I have added a `plugins` key
to `config.markdown.mdex` which is extracted and used to attach MDEx
plugin modules prior to calling `MDEx.to_html!()`.

This _could_ be implemented as a separate key `config.mdex_plugins`, but
since it's part of the markdown converter, this felt more natural to me.
If MDEx or one of its plugins adds a `plugins` option, this will break
that -- but I think it's low probability.